feat: add SwaggerStringAttribute to abstractions
BREAKING-CHANGE: Moves SwaggerStringAttribute to abstractions
